Move out of core package and into ducktape
