+++ bukkit/src/main/java/net/year4000/ducktape/bukkit/module/BukkitModule.java
@@ -2,7 +2,7 @@ package net.year4000.ducktape.bukkit.module;
 
 import net.year4000.ducktape.bukkit.DuckTape;
 import net.year4000.ducktape.bukkit.Settings;
-import net.year4000.ducktape.core.module.AbstractModule;
+import net.year4000.ducktape.module.AbstractModule;
 
 import java.io.File;
