+++ bungee/src/main/java/net/year4000/ducktape/bungee/DuckTape.java
@@ -4,10 +4,10 @@ import com.ewized.utilities.bungee.BungeePlugin;
 import lombok.Getter;
 import net.year4000.ducktape.bungee.module.BungeeModule;
 import net.year4000.ducktape.bungee.module.ModuleManagerListener;
-import net.year4000.ducktape.core.loader.ClassFolderModuleLoader;
-import net.year4000.ducktape.core.loader.ClassModuleLoader;
-import net.year4000.ducktape.core.loader.JarFileModuleLoader;
-import net.year4000.ducktape.core.module.ModuleManager;
+import net.year4000.ducktape.loader.ClassFolderModuleLoader;
+import net.year4000.ducktape.loader.ClassModuleLoader;
+import net.year4000.ducktape.loader.JarFileModuleLoader;
+import net.year4000.ducktape.module.ModuleManager;
 
 public class DuckTape extends BungeePlugin {
     private static DuckTape inst;
