+++ bungee/src/main/java/net/year4000/ducktape/bungee/DuckTapeModule.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.ProxyServer;
 import net.year4000.ducktape.bungee.module.BungeeModule;
-import net.year4000.ducktape.core.module.ModuleInfo;
+import net.year4000.ducktape.module.ModuleInfo;
 
 import java.util.Arrays;
 import java.util.HashSet;
