+++ core/src/main/java/net/year4000/ducktape/api/events/ModuleEnableEvent.java
@@ -4,8 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import net.year4000.ducktape.core.module.AbstractModule;
-import net.year4000.ducktape.core.module.ModuleInfo;
+import net.year4000.ducktape.module.AbstractModule;
+import net.year4000.ducktape.module.ModuleInfo;
 
 @Data
 @EqualsAndHashCode(callSuper = false)
