+++ core/src/main/java/net/year4000/ducktape/loader/ClassFolderModuleLoader.java
@@ -1,6 +1,6 @@
-package net.year4000.ducktape.core.loader;
+package net.year4000.ducktape.loader;
 
-import net.year4000.ducktape.core.module.ModuleManager;
+import net.year4000.ducktape.module.ModuleManager;
 
 import java.io.File;
 import java.net.MalformedURLException;
