+++ core/src/main/java/net/year4000/ducktape/module/ModuleManager.java
@@ -1,4 +1,4 @@
-package net.year4000.ducktape.core.module;
+package net.year4000.ducktape.module;
 
 import com.ewized.utilities.core.util.LogUtil;
 import com.google.common.eventbus.EventBus;
